Use the new feature field
Removes all of the in-extension stuff and the special page, adds a
preference section for beta features and uses that instead.

Also has two new field classes that *should* be in core. But I've put
them in the extension for now because core doesn't want the checkbox
redesign just yet.

Change-Id: Ie2c959760be50d8b8068b1544aa722d55d2a773c