Change the text in the preferences to "trying this feature"
The current text in the preferences page says:
$1 users have enabled this feature.

A Wikipedia editor complained that it is unclear whether
this is the total number of users who ever tried enabling this feature
or the number of users who are trying it now.
I propose to change the text to make it clear that we are talking
about users who are trying it now.

Change-Id: I36d8549d84e6170d34d4048bc5bc74145b81115b