Fix Beta checkboxes not rendering in Opera 12 (all systems)
In theory Opera 12 is able to render SVGs as background images
but does have problems when SVG backgrounds are used in
combination with border-radius. The SVG is not rendered or
disappears when scrolling. I can confirm this on Windows and Ubuntu.

I'm aware that the Beta features are not made to work in Opera 12.
Anyway, most of them do work with no or only small problems.
I really don't see a reason why we should block Opera 12 users from
testing the Beta features. At least not by simply not rendering a
crucial UI element.

Most other SVG images we use don't have this problem because they
are not used in combination with border-radius.

This issue is discussed here:
https://www.mediawiki.org/wiki/Talk:Beta_Features#Checkboxes

Change-Id: Ice592bc823ef10ca7bf014046aba12efa1db79cb