Unset the preference key before adding it
Some extensions have a dual mechanism to enable them, in the classical
framework (GetPreferences) and in the BetaFeatures framework
(GetBetaFeaturePreferences). This ensures the BF description is added at the
end of the $prefs array instead of reusing the place of the previous key,
dirupting the order in the BF tab.

Bug: 71088
Change-Id: Icefb9492872451fa43ac6258e8f6b0302d81ccbd