Fix a failing test
The `hr` tag may not include a space before the closing bracket. Change
the test to be flexible, as we are only concerned about the existence
of the tag and not how it is formatted.

This was broken in I85a5a59fd0955c1a112e8b24b933f0d9e983a156 which
removes a trailing space in a self-closing tag.

Bug: T132362
Change-Id: Id6b5aa692fbee9c4c262d5760dac33351a5a30e7