Do not try to update user counts for anonymous users
This is what this code does:
* $user *can* be anonymous. I dug deep into the code and it looks like
  this is an expected behavior.
* $oldUser will be null for IP users, possibly crashing any code doing
  something with $oldUser->… below.
* But this code is usually not exectued, because the $betaFeatures array
  is empty.
* But the array is not *guaranteed* to be empty. It is filled by all kinds
  of extensions that may or may not check if the user is anonymous.

This error showed up in the Wikibase CI. There are many, many ways to fix
this, but I found this the most obvious one. An other one is a
 if ( $user->isAnon() ) {
     return;
 }
at the top of this static function.

I removed the return because a positive "true" is the default for a long,
long time now for all hooks.

Bug: T155600
Change-Id: I2a09062e60e5abe19017a2bb98f7b22e76efc933