Refactor auto enroll feature to avoid CAS update exceptions
The api module to save options is calling hook 'GetPreferences'
indirectly by getting all valid preferences.
When the api module now do the requested updates that conflicts with the
auto enroll update in BetaFeatures hook. On index.php this does not
happen, because show of Special:Preferences would already trigger the
update.

Bug: T202149
Change-Id: I3c5b6e15877f255aa50af94c3e6c310ff006805b