UserCountsJob does not need a Title
MediaWiki 1.34.0 supports that do not need a Title object, which is the
case for `UpdateBetaFeatureUserCountsJob`. Change it from a `Job` to a
`GenericParameterJob` and update the sole caller to no more passes
`Title::newMainPage()`.

It addresses an edge case which ended up calling User::loadFromSession()
in Setup.php when registering the job.

Bug: T265125
Change-Id: I72c094f0b0527a4fbd14c630fe2eb018333d7083