Use native method to disable checkboxes when "enable most features" is selected
This change breaks the compatibility with mediawiki under 1.38

Preferences controlled by the auto-enroller can't be changed individually when it's on, but don't disable it initially since 'disable-if' logic is designed not to function on initially disabled fields.

This can also fix T292454 since GlobalPreferences treat disabled fields as not globalizable.

Bug: T298300
Bug: T292454
Change-Id: Ie76f0c6782fe53b96f7a71e2c89e6a44556d9146