Use assertSame instead of assertEquals in tests, and a minor fix for passing the test
Since '===' is used almost everywhere, assertEquals is not quite fit for tests.

The 'betafeatures-auto-enroll' preference is not related to HTMLFeatureField, also fix it to work for the test when the options are not saved.
Difference may not be seen in the real use case, since the database would convert true to '1'.

Change-Id: I4028cea7382fbe84a3cb2421c3872a76b05b345d