Handle the htmlform.enhance hook
This didn't fix the root cause of the bug, but is a good step forward.
Because generally, extensions should handle this hook for performance
on Special:Preference.

We now rely on the auto-infuse feature to load the dependency and
infuse before this script. For the root cause, we should investigate
how to handle the dependency, so that extensions don't need to be
aware that their checkbox widgets were changed to toggles.

Bug: T336933
Change-Id: I27655e3d60a9a457e66038bab4666e3a802ebb30