tests: Avoid TestUser, mock User instead
This test doesn't need a real user, it doesn't need any special user
groups, and in fact it works just fine with a mock.

As a follow-up to I4f63beb1, rename BetaFeaturesTestCase to
BetaFeaturesDatabaseTestCase, to make it clear that all tests extending
this class should be in the database group. Remove the 'BetaFeature'
group because it's not a test class.

Change-Id: I5c330094b445740b49e3ca85458d962263f72968